Move DDR systems under src/systems/ddr/
Restructure DDR computer systems into a shared parent folder, mirroring
the commodore/ grouping pattern:
  src/systems/lc80/   → src/systems/ddr/lc80/
  src/systems/z1013/  → src/systems/ddr/z1013/
  src/systems/z9001/  → src/systems/ddr/z9001/
  src/systems/kc85/   → src/systems/ddr/kc85/

Also:
- Update relative include paths (../../ → ../../../) for new depth
- Add missing is_system_ready() / system_ready_ to DDR system headers
- Fix Z1013 RAM_SIZE_64K type: uint16_t → uint32_t (0x10000 overflow)
- Update CMakeLists.txt source paths

+++ src/systems/ddr/z1013/z1013_constants.h
@@ -23,7 +23,7 @@ inline constexpr uint32_t CPU_FREQ_HZ          = 2000000;   // 2 MHz
 // ── Memory ──────────────────────────────────────────────────────────────
 inline constexpr uint16_t RAM_BASE             = 0x0000;
 inline constexpr uint16_t RAM_SIZE_16K         = 0x4000;    // 16 KB standard
-inline constexpr uint16_t RAM_SIZE_64K         = 0x10000;   // 64 KB expanded
+inline constexpr uint32_t RAM_SIZE_64K         = 0x10000;   // 64 KB expanded
 
 inline constexpr uint16_t MONITOR_ROM_BASE     = 0xF000;
 inline constexpr uint16_t MONITOR_ROM_SIZE     = 0x0800;    // 2 KB
